// apps/pwa/src/components/WomenPageLayout/types.d.ts
import { HeaderIconEnums } from './enums';

export type HeaderProps = {
  leftElement1?: keyof typeof HeaderIconEnums;
  leftElement2?: keyof typeof HeaderIconEnums;
  rightElement?: keyof typeof HeaderIconEnums;
  hasScalingAnimation?: boolean;
  rightElementScript?: string;
  style?: React.CSSProperties;
  middleScript?: string;
  className?: string;
};

export type WomenPageLayoutProps = HeaderProps & {
  paddingBottom?: number | string;
  paddingTop?: number | string;
  children: React.ReactNode;
  headerClassName?: string;
};

export type IconsListTypes = {
  [key in HeaderIconEnums]: React.JSX;
};

export type IconComponentMakerProps = (propd: { iconElement: React.JSX; path: string; loadingId: string }) => React.JSX;